Created api folder, split api.js into separate files for auth, coaster, forum, image, social, groups, updated API imports, deleted unused inner project repo
