+++ src/PersonalRanking.jsx
@@ -1,6 +1,6 @@
 import { useState, useContext, useEffect } from "react"
 import { useNavigate, Link } from "react-router-dom"
-import { fetchParks, fetchCoasters, setFavorite } from "./api"
+import { fetchParks, fetchCoasters, setFavorite } from './api/coasterApi'
 
 import { AuthContext } from "./context"
 import { CoasterContext } from "./coasterContext"
